requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.27.1
pydantic==2.8.2
pydantic-settings==2.2.1

motor==3.5.1
pymongo==4.7.2

argon2-cffi==23.1.0
PyJWT==2.9.0
python-multipart==0.0.9

azure-storage-blob==12.20.0
azure-storage-queue==12.9.0

